Fix Download page tests
The previous download tests asserted that a specific date was in the
filename for any given downloaded file. That doesn't really test the
requirement that a file be "recent", and has been failing on the 2011
edition as we had to manually build it recently. In addition, going
forward the 2011 file will be generated on the first Saturday of the
quartely month, which will only be on the first occasionally.